requirements.txt
langgraph
langchain-google-genai>=2.0.0
google-genai>=0.3
tavily-python
streamlit
python-dotenv
pydantic
